Move OptionsReadRepositoryConfigFile tests to tests/unit directory
